refactor: 예약 삭제 쿼리 개선

+++ src/main/java/roomescape/service/admin/AdminReservationService.java
@@ -1,6 +1,6 @@
 package roomescape.service.admin;
 
-import java.util.List;
+import java.util.Optional;
 import org.springframework.stereotype.Service;
 import roomescape.common.TimeUtils;
 import roomescape.domain.member.Member;
@@ -9,13 +9,10 @@ import roomescape.domain.reservation.ReservationDate;
 import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.domain.slot.Slot;
 import roomescape.domain.theme.Theme;
-import roomescape.domain.waiting.Waiting;
-import roomescape.domain.waiting.WaitingWithRank;
 import roomescape.dto.request.reservation.AdminReservationPreservationRequest;
 import roomescape.dto.response.reservation.ReservationPreservationResponse;
 import roomescape.exception.BadRequestException;
 import roomescape.exception.ConflictException;
-import roomescape.exception.RoomescapeException;
 import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
@@ -55,39 +52,29 @@ public class AdminReservationService {
     }
 
     public void remove(final Long reservationId) {
-        if (reservationRepository.existsById(reservationId)) {
-            Reservation reservation = getReservation(reservationId);
-            Slot slot = reservation.getSlot();
-            if (waitingRepository.existsBySlot(slot)) {
-                promotedWaiting(slot);
-            }
+        Optional<Reservation> reservationOpt = reservationRepository.findById(reservationId);
+        if (reservationOpt.isEmpty()) {
+            return;
         }
+        Reservation reservation = reservationOpt.get();
+
+        Slot slot = reservation.getSlot();
+        promotedWaiting(slot);
         reservationRepository.deleteById(reservationId);
     }
 
     private void promotedWaiting(final Slot slot) {
-        List<WaitingWithRank> waitings = waitingRepository.findWaitingsWithRankBySlot(slot);
-        Waiting promotedWaiting = findWaiting(waitings);
-        waitingRepository.deleteById(promotedWaiting.getId());
-        reservationRepository.save(new Reservation(promotedWaiting.getSlot(), promotedWaiting.getMember()));
-    }
-
-    private Waiting findWaiting(final List<WaitingWithRank> waitings) {
-        return waitings.stream()
-                .filter(w -> w.getRank() == 1)
-                .map(WaitingWithRank::getWaiting)
-                .findFirst()
-                .orElseThrow(() -> new RoomescapeException("Server internal exception"));
+        waitingRepository.findTopRankWaitingBySlot(slot.getTheme(), slot.getDate(),
+                        slot.getTime())
+                .ifPresent(waiting -> {
+                    waitingRepository.deleteById(waiting.getId());
+                    reservationRepository.save(new Reservation(waiting.getSlot(), waiting.getMember()));
+                });
     }
 
     private ReservationTime getReservationTime(final Long timeId) {
         return reservationTimeRepository.findById(timeId)
                 .orElseThrow(() -> new IllegalArgumentException("예약 시간이 존재하지 않습니다."));
-    }
-
-    private Reservation getReservation(final Long reservationId) {
-        return reservationRepository.findById(reservationId)
-                .orElseThrow(() -> new IllegalArgumentException("예약이 존재하지 않습니다."));
     }
 
     private Theme getTheme(final Long themeId) {
